# src/main/java/com/fetch_rewards/receipt_processor/exception/ErrorResponse.java
package com.fetch_rewards.receipt_processor.exception;

/**
 * This is a simple class for handling custom errors.
 */
public class ErrorResponse {
    private String message;

    public ErrorResponse(String message) {
        this.message = message;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }
}